Trying to estimate the HMM hyperparameters in an isolated way. Not working yet.
Two important findings:

* There is an inconsistency between the multivariate gaussian pdf from stats and
the multivariate gaussian random sampler from numpy. In the former the covariance
is seen as a covariance matrix. That's why the flas allow_singular was set to True.

* The GuyZ implementation seems to have numerical problems because the non-scaled
version of alpha and beta are being used. To fix this is necessary to implement the
scaled versions. See Bishop for this.